+++ src/php/DataSift/Storyplayer/Cli/CreateStory/Command.php
@@ -117,6 +117,11 @@ class CreateStory_Command extends CliCommand
         $story = <<<EOS
 <?php
 
+use Storyplayer\SPv2\Modules\Asserts;
+use Storyplayer\SPv2\Modules\Checkpoint;
+use Storyplayer\SPv2\Modules\Log;
+use Storyplayer\SPv2\Stories\BuildStory;
+
 EOS;
 
         if (isset($engine->options->basedOn)) {
@@ -132,28 +137,17 @@ EOS;
 //
 // ------------------------------------------------------------------------
 
-\$story = newStoryFor('Top-Level Category')
-         ->inGroup(['Group inside Top-level Category'])
-         ->called('Your story name')
+\$story = BuildStory::newStory();
 EOS;
 
         if (isset($engine->options->basedOn)) {
-            $i = 0;
             foreach ($engine->options->basedOn as $templateClass) {
-                if ($i == 0) {
-                    $story .= "\n         ->basedOn(new " . basename(str_replace('\\', '/', $templateClass)) . ")";
-                }
-                else {
-                    $story .= "\n         ->andBasedOn(new " . basename(str_replace('\\', '/', $templateClass)) . ")";
-                }
-                $i++;
+                $story .= "\n\$story->basedOn(new " . basename(str_replace('\\', '/', $templateClass)) . ");";
             }
         }
 
-        $story .= ";";
         $story .= <<<EOS
 
-\$story->requiresStoryplayerVersion(2);
 
 // ========================================================================
 //
@@ -163,13 +157,26 @@ EOS;
 
 /*
 \$story->addTestSetup(function() {
+    // what are we doing?
+    \$log = Log::usingLog()->startAction("describe what we are doing");
+
     // setup the conditions for this specific test
+    \$checkpoint = Checkpoint::getCheckpoint();
+
+    // all done
+    \$log->endAction();
 });
 */
 
 /*
 \$story->addTestTeardown(function() {
+    // what are we doing?
+    \$log = Log::usingLog()->startAction("describe what we are doing");
+
     // undo anything that you did in addTestSetup()
+
+    // all done
+    \$log->endAction();
 });
 */
 
@@ -181,7 +188,13 @@ EOS;
 
 /*
 \$story->addPreTestPrediction(function() {
+    // what are we doing?
+    \$log = Log::usingLog()->startAction("describe what we are doing");
+
     // if it is okay for your story to fail, detect that here
+
+    // all done
+    \$log->endAction();
 });
 */
 
@@ -193,23 +206,39 @@ EOS;
 
 /*
 \$story->addPreTestInspection(function() {
+    // what are we doing?
+    \$log = Log::usingLog()->startAction("describe what we are doing");
+
     // get the checkpoint - we're going to store data in here
-    \$checkpoint = getCheckpoint();
+    \$checkpoint = Checkpoint::getCheckpoint();
 
     // store any data that your story is about to change, so that you
     // can do a before and after comparison
+
+    // all done
+    \$log->endAction();
 });
 */
 
 // ========================================================================
 //
-// POSSIBLE ACTION(S)
+// ACTIONS
 //
 // ------------------------------------------------------------------------
 
 /*
 \$story->addAction(function() {
+    // what are we doing?
+    \$log = Log::usingLog()->startAction("describe what we are doing");
+
+    // use the checkpoint to store any data collected during the action
+    // this data will be examined in the postTestInspection phase
+    \$checkpoint = Checkpoint::getCheckpoint();
+
     // this is where you perform the steps of your user story
+
+    // all done
+    \$log->endAction();
 });
 */
 
@@ -220,12 +249,18 @@ EOS;
 // ------------------------------------------------------------------------
 
 \$story->addPostTestInspection(function() {
+    // what are we doing?
+    \$log = Log::usingLog()->startAction("describe what we are doing");
+
     // the information to guide our checks is in the checkpoint
-    \$checkpoint = getCheckpoint();
+    \$checkpoint = Checkpoint::getCheckpoint();
 
     // gather new data, and make sure that your action actually changed
     // something. never assume that the action worked just because it
     // completed to the end with no errors or exceptions!
+
+    // all done
+    \$log->endAction();
 });
 
 EOS;